699: separate component for help icon
