+++ src/Components/HelpBubbleIcon/helpBubbleIcon.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import HelpBubble from '../Assets/helpBubble.svg';
+import HelpBubble from '../../Assets/helpBubble.svg';
 
-export function HelpBubbleIcon({ className }: { className?: string}) {
+export function HelpBubbleIcon({ className}: { className?: string}) {
   return (
       <img className={className} alt="MyFriendBen Help Bubble Icon" src={HelpBubble}  
       style={{ height: '20px', width: '20px' }}
