+++ src/Components/HouseholdDataBlock/HouseholdDataBlock.js
@@ -23,8 +23,8 @@ import {
 } from '../../Assets/validationFunctions.tsx';
 import { getStepNumber } from '../../Assets/stepDirectory';
 import { Context } from '../Wrapper/Wrapper.tsx';
-import { HelpBubbleIcon } from '../../Assets/helpBubbleIcon.tsx';
 import './HouseholdDataBlock.css';
+import HelpButton from '../HelpBubbleIcon/HelpButton.tsx';
 
 const HouseholdDataBlock = ({ handleHouseholdDataSubmit }) => {
   const { formData } = useContext(Context);
@@ -371,18 +371,12 @@ const HouseholdDataBlock = ({ handleHouseholdDataSubmit }) => {
         <Box className="section">
           <h2 className="question-label">
             <FormattedMessage id={formattedMsgId} defaultMessage={formattedMsgDefaultMsg} />
-            <IconButton onClick={handleClick} >
-            <HelpBubbleIcon/>
-          </IconButton>
-        </h2>
-          <p className="question-description help-text">
-            {showHelpText && (
-            <FormattedMessage
-                id="householdDataBlock.createIncomeRadioQuestion-questionDescription"
-                defaultMessage="This includes money from jobs, alimony, investments, or gifts. Income is the money earned or received before deducting taxes"
-              />
-            )}
-        </p>
+            <HelpButton
+              isVisible={true}
+              helpText="This includes money from jobs, alimony, investments, or gifts. Income is the money earned or received before deducting taxes"
+              helpId="householdDataBlock.createIncomeRadioQuestion-questionDescription"
+            ></HelpButton>
+          </h2>
           <HHDataRadiofield componentDetails={radiofieldProps} memberData={memberData} setMemberData={setMemberData} />
         </Box>
       </Box>
